backend/requirements.txt
fastapi
uvicorn
openai-whisper
moviepy
torch
pydantic
spacy